Rename UnitTestingWidget to UnitTestWidget
